Separate monaco from main app and split up Codeeditor between embed and app, fix double linter issue
