+++ .gitignore
@@ -21,6 +21,7 @@ npm-debug.log
 yarn-error.log
 .idea
 jest
+stats.json
 
 lerna-debug.log
