--- packages/app/src/app/components/CodeEditor/Monaco/workers/linter/vue/default-config.js
@@ -1,50 +1,0 @@
-export default {
-  parser: 'vue-eslint-parser',
-  parserOptions: Object.freeze({
-    parser: 'babel-eslint',
-    ecmaVersion: 2015,
-    sourceType: 'module',
-    ecmaFeatures: Object.freeze({
-      jsx: true,
-      experimentalObjectRestSpread: true,
-    }),
-  }),
-  env: {
-    browser: true,
-    es6: true,
-  },
-  plugins: ['vue'],
-  rules: {
-    'vue/comment-directive': 'error',
-    'vue/jsx-uses-vars': 'error',
-    'vue/no-async-in-computed-properties': 'error',
-    'vue/no-dupe-keys': 'error',
-    'vue/no-duplicate-attributes': 'error',
-    'vue/no-parsing-error': 'error',
-    'vue/no-reserved-keys': 'error',
-    'vue/no-shared-component-data': 'error',
-    'vue/no-side-effects-in-computed-properties': 'error',
-    'vue/no-template-key': 'error',
-    'vue/no-textarea-mustache': 'error',
-    'vue/no-unused-vars': 'error',
-    'vue/require-component-is': 'error',
-    'vue/require-render-return': 'error',
-    'vue/require-v-for-key': 'error',
-    'vue/require-valid-default-prop': 'error',
-    'vue/return-in-computed-property': 'error',
-    'vue/valid-template-root': 'error',
-    'vue/valid-v-bind': 'error',
-    'vue/valid-v-cloak': 'error',
-    'vue/valid-v-else-if': 'error',
-    'vue/valid-v-else': 'error',
-    'vue/valid-v-for': 'error',
-    'vue/valid-v-html': 'error',
-    'vue/valid-v-if': 'error',
-    'vue/valid-v-model': 'error',
-    'vue/valid-v-on': 'error',
-    'vue/valid-v-once': 'error',
-    'vue/valid-v-pre': 'error',
-    'vue/valid-v-show': 'error',
-    'vue/valid-v-text': 'error',
-  },
-};
