+++ packages/app/src/app/utils/monaco-index-converter.js
