+++ packages/app/src/embed/components/Content/Monaco/define-theme.js
