+++ packages/app/src/embed/components/Content/Monaco/grammars/configure-tokenizer.js
