+++ packages/app/src/embed/components/Content/Monaco/grammars/tmGrammars/TypeScriptReact.tmLanguage
