+++ packages/app/src/embed/components/Content/Monaco/mode.ts
