+++ packages/app/src/embed/components/Content/Monaco/settings.js
